refactor: resturcture packages to ddd modules
