+++ src/main/java/com/moayo/moayobackend/post/entity/Category.java
@@ -1,4 +1,4 @@
-package com.moayo.moayobackend.entity;
+package com.moayo.moayobackend.post.entity;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
