+++ src/main/java/com/moayo/moayobackend/post/service/PostService.java
@@ -1,9 +1,9 @@
-package com.moayo.moayobackend.service;
+package com.moayo.moayobackend.post.service;
 
-import com.moayo.moayobackend.entity.Category;
-import com.moayo.moayobackend.entity.Post;
-import com.moayo.moayobackend.dto.PostResponseDto;
-import com.moayo.moayobackend.repository.PostRepository;
+import com.moayo.moayobackend.post.entity.Category;
+import com.moayo.moayobackend.post.entity.Post;
+import com.moayo.moayobackend.post.dto.PostResponseDto;
+import com.moayo.moayobackend.post.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
